requirements.txt
amqp==5.2.0
asgiref==3.7.2
awsebcli==3.20.10
billiard==4.2.0
boto3==1.27.1
botocore==1.30.1
celery==5.3.6
cement==2.8.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2.9
django-crispy-forms==2.1
django-debug-toolbar==3.6.0
django-filter==23.5
django-redis==5.4.0
django-redis-cache==3.0.1
django-ses==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filelock==3.13.1
idna==3.6
jmespath==1.0.1
kombu==5.3.5
oauthlib==3.2.2
pathspec==0.10.1
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pypiwin32==223
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
semantic-version==2.8.5
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.1.9
